// src/interface.ts
interface ContestResult {
  IsRated: boolean;
  Place: number;
  OldRating: number;
  NewRating: number;
  Performance: number;
  InnerPerformance: number;
  ContestScreenName: string;
  ContestName: string;
  ContestNameEn: string;
  EndTime: string;
}

export interface AtCoderContestInfo {
  username: string;
  contestResult: ContestResult[];
}